backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.5.3
httpx==0.27.2
zulip>=0.9.1
requests>=2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
passlib[bcrypt]==1.7.4
bcrypt<5.0.0




